requirements.txt
beautifulsoup4
requests
lxml
pandas
psycopg2-binary
sqlalchemy